--- a/onnxruntime/contrib_ops/webgpu/quantization/matmul_nbits.h
+++ b/onnxruntime/contrib_ops/webgpu/quantization/matmul_nbits.h
@@ -45,20 +45,6 @@ class MatMulNBitsWithLargeMProgram final : public Program<MatMulNBitsWithLargeMP
   bool has_zero_points_;
 };
 
-class MatMulNBitsProgramPrefill final : public Program<MatMulNBitsProgramPrefill> {
- public:
-  MatMulNBitsProgramPrefill() : Program{"MatMulNBitsPrefill"} {
-  }
-
-  Status GenerateShaderCode(ShaderHelper& sh) const override;
-  WEBGPU_PROGRAM_DEFINE_UNIFORM_VARIABLES(
-      {"M", ProgramUniformVariableDataType::Uint32},
-      {"N", ProgramUniformVariableDataType::Uint32},
-      {"K", ProgramUniformVariableDataType::Uint32},
-      {"K4", ProgramUniformVariableDataType::Uint32},
-      {"K8", ProgramUniformVariableDataType::Uint32});
-};
-
 class MatMulNBits final : public WebGpuKernel {
  public:
   MatMulNBits(const OpKernelInfo& info) : WebGpuKernel(info) {