@qjia7 2024-12-13
@sushraja-msft @sushanthr Currently, I only test it on my laptop with dual GPUs. You can find the data in above description message. Please help verify it in your side to see if we can see similar result since our gpus and benchmarks are not same. 

cc @guschmue @fs-eire This PR still needs to be further refactored to reduce some duplicated codes. Now it's just for verification.