@sushraja-msft 2024-12-13
> @sushraja-msft @sushanthr Currently, I only test it on my laptop with dual GPUs. You can find the data in above description message. Please help verify it in your side to see if we can see similar result since our gpus and benchmarks are not same.
> 
> cc @guschmue @fs-eire This PR still needs to be further refactored to reduce some duplicated codes. Now it's just for verification.

Ran your change on my intel Xe laptop, this is faster than mine 👏. 55tk's vs 44tk's in mine. 
We should land yours and its okay to remove my implementation @guschmue @qjia7 

```
C:\model_benchmark>model_benchmark.exe -i C:\Phi-3.5-mini-instruct-onnx-web\Phi-3.5-mini-instruct-onnx-web -l 500
Batch size: 1, prompt tokens: 501, tokens to generate: 128
Prompt processing (time to first token):
        avg (us):       9.10299e+06
        avg (tokens/s): 55.0369                                <<<<
        p50 (us):       9.09658e+06
        stddev (us):    13042.6
        n:              5 * 501 token(s)
Token generation:
        avg (us):       79482.3
        avg (tokens/s): 12.5814
        p50 (us):       79505.4
        stddev (us):    2280.06
        n:              635 * 1 token(s)
Token sampling:
        avg (us):       18.0841
        avg (tokens/s): 55297.3
        p50 (us):       14.4
        stddev (us):    24.9088
        n:              640 * 1 token(s)
E2E generation (entire generation loop):
        avg (ms):       19199.6
        p50 (ms):       19200.1
        stddev (ms):    20.0724
        n:              5
Peak working set size (bytes): 5470642176
WebGPU device lost (2): Device was destroyed.```