@guschmue 2024-12-14
very cool JiaJia, I can run it on a bunch of machines